requirements.txt
flask
flask_cors
pymongo
pandas
openpyxl
PyJWT
Werkzeug
gunicorn
requests
dotenv
pymongo
sqlalchemy